# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Urja Ops Portal REST API Wrapper",
        "version": "1.0.0",
        "description": "A read-only REST API wrapper built over the reverse-engineered Urja Ops legacy portal. Exposes meters, search, meter details, consumption, and network hierarchy endpoints."
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Local development server"
        }
    ],
...
    "paths": {
...
        "/meters": {
            "get": {
                "summary": "List all meters",
                "description": "Returns a paginated list of meters from the legacy portal catalog.",
                "tags": [
                    "Meters"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1
                        },
                        "description": "Page number for pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of meters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MeterSummary"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Portal unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "summary": "Search meters",
                "description": "Search the meter catalog by meter ID, serial number, make, or other attributes.",
                "tags": [
                    "Meters"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "q",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search query (meter ID, serial number, make, etc.)",
                        "example": "Genus"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1
                        },
                        "description": "Page number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results matching query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MeterSummary"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Portal unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/meters/{id}": {
            "get": {
                "summary": "Get meter details",
                "description": "Returns full meter metadata including serial number, make, install type, geo coordinates, and network hierarchy.",
                "tags": [
                    "Meters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Meter ID",
                        "example": "J100008"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Meter details with coordinates and hierarchy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MeterDetail"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid meter ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Portal unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/meters/{id}/consumption": {
            "get": {
                "summary": "Get consumption readings",
                "description": "Returns historical energy consumption readings with ISO timestamps, kWh, kVAh, and voltage values.",
                "tags": [
                    "Consumption"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Meter ID",
                        "example": "J100008"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of consumption readings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ConsumptionReading"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid meter ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Portal unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/network": {
            "get": {
                "summary": "Get distribution network tree",
                "description": "Returns the full hierarchical tree of distribution transformers (DTs) with their child meters.",
                "tags": [
                    "Network"
                ],
                "responses": {
                    "200": {
                        "description": "Network tree of transformers with grouped meters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NetworkNode"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Portal unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "MeterSummary": {
                "type": "object",
                "properties": {
                    "meterId": {
                        "type": "string",
                        "example": "J100008"
                    },
                    "serialNo": {
                        "type": "string",
                        "example": "SE63900"
                    },
                    "make": {
                        "type": "string",
                        "example": "Genus"
                    },
                    "phaseType": {
                        "type": "string",
                        "example": "single"
                    },
                    "installStatus": {
                        "type": "string",
                        "example": "Installed"
                    },
                    "dtCode": {
                        "type": "string",
                        "example": "DT-009"
                    }
                }
            },
            "MeterDetail": {
                "type": "object",
                "properties": {
                    "meterId": {
                        "type": "string",
                        "example": "J100008"
                    },
                    "serialNo": {
                        "type": "string",
                        "example": "SE63900"
                    },
                    "make": {
                        "type": "string",
                        "example": "Genus"
                    },
                    "phaseType": {
                        "type": "string",
                        "example": "single"
                    },
                    "installStatus": {
                        "type": "string",
                        "example": "Installed"
                    },
                    "installType": {
                        "type": "string",
                        "example": "Whole Current"
                    },
                    "coordinates": {
                        "type": "object",
                        "properties": {
                            "latitude": {
                                "type": "number",
                                "example": 26.899010
                            },
                            "longitude": {
                                "type": "number",
                                "example": 75.840067
                            }
                        }
                    },
                    "hierarchy": {
                        "type": "object",
                        "properties": {
                            "zone": {
                                "type": "string",
                                "example": "Jaipur Zone"
                            },
                            "circle": {
                                "type": "string",
                                "example": "Jaipur City Circle"
                            },
                            "division": {
                                "type": "string"
                            },
                            "subdivision": {
                                "type": "string"
                            },
                            "substation": {
                                "type": "string"
                            },
                            "feeder": {
                                "type": "string"
                            },
                            "dt": {
                                "type": "string",
                                "example": "Sikar Road DT 9"
                            }
                        }
                    }
                }
            },
            "ConsumptionReading": {
                "type": "object",
                "properties": {
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-23T23:30:00.000Z"
                    },
                    "kwh": {
                        "type": "number",
                        "example": 27623.95
                    },
                    "kvah": {
                        "type": "number",
                        "example": 29833.87
                    },
                    "voltage": {
                        "type": "number",
                        "example": 227
                    }
                }
            },
            "NetworkNode": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "DT-001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Malviya Nagar DT 1"
                    },
                    "feederCode": {
                        "type": "string",
                        "example": "F-001"
                    },
                    "capacityKva": {
                        "type": "number",
                        "example": 100
                    },
                    "meters": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "meterId": {
                                    "type": "string"
                                },
                                "serialNo": {
                                    "type": "string"
                                },
                                "make": {
                                    "type": "string"
                                },
                                "phaseType": {
                                    "type": "string"
                                },
                                "installStatus": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "PaginationMeta": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "pageSize": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 403
                    }
                }
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    }
}